Load data

+++ linkedin/profile/scrape_profile.py
@@ -17,10 +17,24 @@ with open("result_dict.json", "r") as fp:
 
 count = 0
 for url in result:
-	driver.get(url)
 	trim_url = url.split('/')[-2]
+	if os.path.exists("../html/" + trim_url + ".html"):
+		print ("duplicated...")
+		continue
+
+
+	driver.get(url)
+	time.sleep(5)
+	driver.execute_script("window.scrollTo(0, document.body.scrollHeight*0.5)")
+	time.sleep(5)
+	driver.execute_script("window.scrollTo(0, document.body.scrollHeight)")
+	time.sleep(5)
+	
+
 	with open("../html/" + trim_url + ".html", 'w') as fp:
 		fp.write(driver.page_source)
 
+
 	print (count)
-	++count
+	time.sleep(2)
+	count += 1
